docker/spark/app/requirements.txt
redis
python-dotenv
pyspark

